Move ProblemDetails to solid module, remove ClientHttpException
